Split up editor tests into multiple files
