+++ test/e2e/tests/__expected-screenshots__/editor/editor.pictures.spec.ts/Editor-pictures-First-picture-and-caption-is-present-1-chromium.png
