+++ test/e2e/utils/project-utils.ts
@@ -1,9 +1,14 @@
 import { APIRequestContext, TestInfo } from "@playwright/test";
-import { Project, UserDetails } from "./types";
+import { Project, TestProject, UserDetails } from "./types";
 import { getTestControl, TestControlService } from './test-control-api';
 import { TestFile } from "./path-utils";
+import { users, entries } from "../constants";
 
-const toProjectCode = (name: string): string => name.toLowerCase().replace(/[ \.]/g, '_');
+const toProjectCode = (name: string): string => {
+  name = name.toLowerCase().replace(/[ \.]/g, '_');
+  const slash = name.indexOf('/');
+  return name.substring(slash + 1);
+}
 
 export const toProject = (name: string, id?: string): Project => ({
   name,
@@ -64,5 +69,20 @@ export class ProjectTestService {
   addLexEntry(projectCode: string | Project, data: any) {
     if (data.id == null) data.id = '';
     return this.call('add_lexical_entry', [getCode(projectCode), data]) as Promise<string>;
+  }
+
+  async createDefaultProject(testInfo: TestInfo): Promise<TestProject> {
+    const project = await this.initTestProject(testInfo.titlePath[0], undefined, users.manager, [users.member]);
+    await this.addUserToProject(project, users.observer, "observer");
+    await this.addWritingSystemToProject(project, 'th-fonipa', 'tipa');
+    await this.addWritingSystemToProject(project, 'th-Zxxx-x-audio', 'taud');
+    await this.addPictureFileToProject(project, entries.entry1.senses[0].pictures[0].fileName);
+    await this.addAudioVisualFileToProject(project, entries.entry1.lexeme['th-Zxxx-x-audio'].value);
+    const projectEntries = await Promise.all([
+      this.addLexEntry(project.code, entries.entry1),
+      this.addLexEntry(project.code, entries.entry2),
+      this.addLexEntry(project.code, entries.multipleMeaningEntry),
+    ]);
+    return { project: () => project, entryIds: () => projectEntries };
   }
 }
